Git差分コメント機能の追加とCLIの拡張
- 'git diff-comment' サブコマンドを追加し、Git差分からコミットメッセージを生成する機能を実装
- CLIの構造を改善し、'git' コマンドカテゴリーを追加
- シェル補完スクリプトを更新し、新しいコマンドとオプションに対応
- 'copy-git-diff' オプションを setup.sh から削除（機能が不要になったため）
- パッケージ構造を整理し、llm パッケージを再構成
